# getJsonObject: improve performance for new version of getJsonObject