--- a/src/main/java/com/nvidia/spark/rapids/jni/JSONUtils.java
+++ b/src/main/java/com/nvidia/spark/rapids/jni/JSONUtils.java
@@ -23,7 +23,8 @@ public class JSONUtils {
     NativeDepsLoader.loadNativeDeps();
   }
 
-  public static final int MAX_PATH_DEPTH = 32;
+  // Keep the same with `max_path_depth` in `get_json_object.cu'
+  public static final int MAX_PATH_DEPTH = 8;
 
   public enum PathInstructionType {
     SUBSCRIPT,