--- a/src/main/java/com/nvidia/spark/rapids/jni/JSONUtils.java
+++ b/src/main/java/com/nvidia/spark/rapids/jni/JSONUtils.java
@@ -24,7 +24,7 @@ public class JSONUtils {
   }
 
   // Keep the same with `max_path_depth` in `get_json_object.cu'
-  public static final int MAX_PATH_DEPTH = 8;
+  public static final int MAX_PATH_DEPTH = 16;
 
   public enum PathInstructionType {
     SUBSCRIPT,