… src/main/cpp/src/json_parser.cuh
… src/test/java/com/nvidia/spark/rapids/jni/GetJsonObjectTest.java