closes #1907

Improve performance for new version of getJsonObject.
Fix a redefined variable

### perf test
Tested ByteDance data.
The first line is cuDF impl.

|After commit|time(1st,2nd,3rd)| time / original cuDF impl (less is better)|
|:-|:-|:-|
|original cuDF impl, branch 24.02| 2183 ms, 778 ms, 756 ms| base 1.0|
|new version of get-json-obj in branch 24.04| 21230 ms, 19887 ms, 20150 ms|16.482916330374|
|Refactor: Move device code in get_json_object to cu or cuh files|21278 ms, 19917 ms, 20366 ms|16.5620123755717|
|Refactor/simplify json generator|6113 ms, 4734 ms, 4690 ms|4.17998385794996|
|Remove purge non-empty nulls step|6064 ms, 4747 ms, 4753 ms|4.18724778046812|
|Refactor josn parser: remove useless variables|6013 ms, 4687 ms, 4752 ms|4.15711595372612|
|Use 64 bits to store JSON nest depth context to save memory|6086 ms, 4695 ms, 4709 ms|4.16733925208502|
|Refine push/pop/peek logic|2850 ms, 1570 ms, 1595 ms|1.6182405165456|
|Refactor: use less functions; change max path length from 32 to 8|2716, 1584, 1515|1.56443368307775|


Signed-off-by: Chong Gao <res_life@163.com>