@res-life 2024-04-07
Original cuDF uses time about 700ms when testing ByteDance data.

|commit|time|
|:-:|:-:|
|init new version|15s|
|json_nest_depth = 8; max_path=4|5s|
|use bits in int64 to save json nest depth; json_nest_depth = 64; max_path=8|5s|
|Refactor: reduce push/pop; json_nest_depth = 64; max_path=8 |1.7s|
|Refactor: reduce construct context; json_nest_depth = 64; max_path=8 |1.6s|
|Refactor generator |1.6s|
|allocate uvector<context> |avg 700ms|

After some commits, we can get 700ms as original cuDF impl.

Note: currently meets a error in QA's env:
```
ai.rapids.cudf.CudaFatalException: parallel_for: failed to synchronize: cudaErrorIllegalAddress: an illegal memory access was encountered
```