@res-life 2024-04-09
Please refer to the last commit, it fix nvbug: Fix nvbug: get-json-obj get incorrect result on some GPUs(H100, V100).
Seems it's related to `thrust::tie`, on some GPU(H100 V100) CUDA(12.x), tie generate whong result(empty output). And on T4, does not occur. It's weird issue.
After changed to tie a structure instead of `cudf::size_type`, then generate correct answer.
Currently do not know why this fix works.