@res-life 2024-04-10
>Please try reverting get_json_object.cu in the latest commit and apply this:

Done, with a little change:

```
    } else {
      // valid JSON length is always greater than 0
      // if `str` size len is zero, output len is 0 and `is_valid` is false
      if (out_buf == nullptr) { d_sizes[tid] = 0; }
    }
```