@ttnghia 2024-04-11
Performance change after this PR, compared using https://github.com/NVIDIA/spark-rapids-jni/pull/1952:
```
|  size_bytes  |  max_depth  |   Ref Time |   Ref Noise |   Cmp Time |   Cmp Noise |           Diff |   %Diff |  Status  |
|--------------|-------------|------------|-------------|------------|-------------|----------------|---------|----------|
|   1000000    |      2      |   6.654 ms |      10.40% |   5.576 ms |      11.35% |   -1078.003 us | -16.20% |   FAIL   |
|   10000000   |      2      |   7.373 ms |       8.92% |   6.230 ms |      11.22% |   -1143.858 us | -15.51% |   FAIL   |
|  100000000   |      2      | 145.607 ms |       1.81% | 131.515 ms |       1.96% |  -14092.123 us |  -9.68% |   FAIL   |
|  1000000000  |      2      |    1.745 s |       0.67% |    1.558 s |       1.39% | -186955.541 us | -10.72% |   FAIL   |
|   1000000    |      4      |   7.300 ms |       8.87% |   5.836 ms |      12.70% |   -1463.903 us | -20.05% |   FAIL   |
|   10000000   |      4      |   8.247 ms |       7.03% |   6.414 ms |       8.63% |   -1833.158 us | -22.23% |   FAIL   |
|  100000000   |      4      | 159.210 ms |       0.98% | 133.165 ms |       1.58% |  -26044.341 us | -16.36% |   FAIL   |
|  1000000000  |      4      |    1.872 s |       0.92% |    1.518 s |       1.00% | -354327.340 us | -18.93% |   FAIL   |
|   1000000    |      6      |   7.639 ms |       9.03% |   5.905 ms |      10.32% |   -1734.084 us | -22.70% |   FAIL   |
|   10000000   |      6      |   8.646 ms |       8.02% |   6.346 ms |       7.87% |   -2300.449 us | -26.61% |   FAIL   |
|  100000000   |      6      | 165.838 ms |       1.92% | 131.152 ms |       1.16% |  -34686.127 us | -20.92% |   FAIL   |
|  1000000000  |      6      |    1.925 s |       1.07% |    1.581 s |       0.68% | -344136.035 us | -17.87% |   FAIL   |
|   1000000    |      8      |   8.109 ms |       6.55% |   6.090 ms |      10.00% |   -2019.178 us | -24.90% |   FAIL   |
|   10000000   |      8      |   9.414 ms |       5.90% |   6.598 ms |       8.01% |   -2815.621 us | -29.91% |   FAIL   |
|  100000000   |      8      | 177.391 ms |       1.20% | 134.754 ms |       1.20% |  -42636.822 us | -24.04% |   FAIL   |
|  1000000000  |      8      |    2.034 s |       0.54% |    1.576 s |       0.71% | -457780.972 us | -22.51% |   FAIL   |
```